+++ recipes/configs/dev/llama3/8B_dora_fsdp2.yaml
@@ -11,7 +11,7 @@
 # You can add specific overrides through the command line. For example
 # to override the checkpointer directory while launching training
 # you can run:
-#   tune run --nnodes 1 --nproc_per_node 2 lora_finetune_fsdp2 --config llama3/8B_dora checkpointer.checkpoint_dir=<YOUR_CHECKPOINT_DIR>
+#   tune run --nnodes 1 --nproc_per_node 8 lora_finetune_fsdp2 --config llama3/8B_dora checkpointer.checkpoint_dir=<YOUR_CHECKPOINT_DIR>
 
 
 # Model Arguments
@@ -61,7 +61,7 @@ loss:
 
 # Training
 epochs: 1
-max_steps_per_epoch: 50
+max_steps_per_epoch: 1
 gradient_accumulation_steps: 1
 
 # Logging
